Decouple tons of code
