+++ src-tauri/src/config/app_config.rs
@@ -1,6 +1,5 @@
 use config::{Config, ConfigError, File};
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
