+++ src-tauri/src/tray_menu/mod.rs
@@ -1,2 +1,26 @@
 pub mod menu;
 pub mod menu_component;
+
+#[derive(Debug)]
+pub enum MenuId {
+    Quit,
+    Test,
+    Hide,
+    Show,
+    SetClipboard,
+    GetClipboard,
+}
+
+impl MenuId {
+    pub fn parse_menu_id(id: &str) -> Result<MenuId, String> {
+        match id {
+            "quit" => Ok(MenuId::Quit),
+            "test" => Ok(MenuId::Test),
+            "hide" => Ok(MenuId::Hide),
+            "show" => Ok(MenuId::Show),
+            "set_clipboard" => Ok(MenuId::SetClipboard),
+            "get_clipboard" => Ok(MenuId::GetClipboard),
+            _ => Err(format!("Unknown menu id: {}", id)),
+        }
+    }
+}
